Create README.md
The full overview of all learning material in the course is found at https://www.math.ntnu.no/emner/TMA4268/2019v/TMA4268overview.html

The textbook for the course was James et al: Introduction to statistical learning, plus an extra module on artificial neural networks. 
This repo contains my Rmd code for the modules that I have made (not chapters 6,7 and 10 of the ISLR).

The lisence is CC-BY-NC-ND.

Please create an issue if you find errors or inconsitencies.